requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
multipledispatch==0.6.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
tornado==6.1
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
